requirements.txt
altair==4.1.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.5
boto==2.49.0
boto3==1.13.11
botocore==1.16.11
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
contractions==0.0.24
-e git+https://github.com/panas89/coronaBreakSuck2020.git@c648d3dc286c873bc525d76cb18ba826be64194e#egg=covid
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
feedfinder2==0.0.4
feedparser==5.2.1
flake8==3.8.3
funcy==1.14
future==0.18.2
gensim==3.8.3
geograpy3 @ git+git://github.com/jmbielec/geograpy3.git@21b85924f262c7c36f84ac15fff814001746d547
geotext @ https://github.com/elyase/geotext/archive/master.zip
idna==2.9
importlib-metadata==1.6.0
inflect==4.1.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
jellyfish==0.8.2
jieba3k==0.35.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
lxml==4.5.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.3.0
nbconvert==5.6.1
nbformat==5.0.6
newspaper3k==0.2.8
nltk==3.4.5
notebook==6.0.3
numexpr==2.7.1
numpy==1.18.2
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pkg-resources==0.0.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
py==1.8.1
pyahocorasick==1.4.0
pycodestyle==2.6.0
pycountry==19.8.18
pyflakes==2.2.0
Pygments==2.6.1
pyLDAvis==2.1.2
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.2
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
requests==2.23.0
requests-file==1.5.1
s3transfer==0.3.3
scikit-learn==0.23.0
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
smart-open==2.0.0
soupsieve==2.0
terminado==0.8.3
testpath==0.4.4
textsearch==0.0.17
threadpoolctl==2.0.0
tinysegmenter==0.3
tldextract==2.2.2
toolz==0.10.0
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
Unidecode==1.1.1
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
